third_party/blink/web_tests/virtual/css-sizing-keyword-animation/external/wpt/css/css-sizing/animation/min-width-interpolation-expected.txt

This is a testharness.js-based test.
Found 210 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (-0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (0) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (-0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (0) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (-0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (0) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Animations: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (-0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (0) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (0.3) should be [initial]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] Web Animations: property <min-width> from [initial] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (-0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (0) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (-0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (0) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (-0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (0) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Animations: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (-0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (0) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (0.3) should be [unset]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] Web Animations: property <min-width> from [unset] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (-0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (0) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (-0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (0) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (-0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0.3) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (-0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (0) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] CSS Animations: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (-0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , - 6px + size * 1.3 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (0) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 0px + size * 1 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (0.3) should be [auto]
  assert_equals: expected "0px " but got "calc - size ( auto , 6px + size * 0.7 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (0.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 10px + size * 0.5 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (0.6) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 12px + size * 0.4 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (1) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 20px + size * 0 ) "
[FAIL] Web Animations: property <min-width> from [auto] to [20px] at (1.5) should be [20px]
  assert_equals: expected "20px " but got "calc - size ( auto , 30px + size * - 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (-0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (0) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (-0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (0) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [0px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [fit-content] to neutral at (-0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , - 3px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [fit-content] to neutral at (0) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from [fit-content] to neutral at (0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 3px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [fit-content] to neutral at (-0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , - 3px + size * 1.3 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [fit-content] to neutral at (0) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from [fit-content] to neutral at (0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 3px + size * 0.7 ) "
[FAIL] CSS Animations: property <min-width> from [fit-content] to neutral at (-0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , - 3px + 0px + size * 1.3 ) "
[FAIL] CSS Animations: property <min-width> from [fit-content] to neutral at (0) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-width> from [fit-content] to neutral at (0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 3px + 0px + size * 0.7 ) "
[FAIL] Web Animations: property <min-width> from [fit-content] to neutral at (-0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , - 3px + 0px + size * 1.3 ) "
[FAIL] Web Animations: property <min-width> from [fit-content] to neutral at (0) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <min-width> from [fit-content] to neutral at (0.3) should be [fit-content]
  assert_equals: expected "fit - content " but got "calc - size ( fit - content , 3px + 0px + size * 0.7 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (-0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 13px + size * - 0.3 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (0) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 10px + size * 0 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 7px + size * 0.3 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (-0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 13px + size * - 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 10px + size * 0 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0.3) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 7px + size * 0.3 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + size * 0.5 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + size * 0.6 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + size * 1 ) "
[FAIL] CSS Transitions with transition: all: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + size * 1.5 ) "
[FAIL] CSS Animations: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + 0px + size * 0.5 ) "
[FAIL] CSS Animations: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + 0px + size * 0.6 ) "
[FAIL] CSS Animations: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + 0px + size * 1.5 ) "
[FAIL] Web Animations: property <min-width> from neutral to [max-content] at (0.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 5px + 0px + size * 0.5 ) "
[FAIL] Web Animations: property <min-width> from neutral to [max-content] at (0.6) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 4px + 0px + size * 0.6 ) "
[FAIL] Web Animations: property <min-width> from neutral to [max-content] at (1) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <min-width> from neutral to [max-content] at (1.5) should be [max-content]
  assert_equals: expected "max - content " but got "calc - size ( max - content , - 5px + 0px + size * 1.5 ) "
Harness: the test ran to completion.
